Re-enable combined restriction functionality (#86)
* Re-enable combined restriction functionality

Re-enables the functionality that used to be encoded in the Combined restriction type. To avoid circular type references and to simplify the API this has instead been achieved by abstracting out the way restricitons are combined. Previously top-level we assumed that restrictions should be 'AND'ed together. Now an arbitrary binary expression tree is used  written in postfix notation.

Also added are a Fail restriction type which always fails to pass. This is useful both for testing and for doing certain unary operations as the tree only excepts binary operations.

There is some opinionation as to what the defaul should be for the Process program. This change maintains current behaviour that the program will always succeed.

* Version bump

* Cargo fmt

* Update readme

* Fix versions

* Runtime spec version