Feature/restrictions compilation (#146)
Adds restriciton compilation to `dscp-lang`

Majority of the changes here are to add the `build` command to `dscp-lang`. There are some additional more general changes to the runtime that we're needed to get this working. Notably:
1. New restriction type `FixedInputMetadataValueType`
2. Abstraction of types from `dscp-node-runtime` to a subpackage `dscp-runtime-types`. This is to prevent `dscp-lang` becoming dependent on the very heavy runtime
3. Increase max program length to 501 (inc. weight change)

The compiler changes themselvers are fairly complex and relatviely poorly tested. I would like to improve test coverage and refactoring (esp. tools/lang/src/compiler/condition_transform.rs) but I think this can wait for now.